pyproject.toml
[project]
name = "bt-miner-skills"
version = "0.1.0"
description = "Agent-driven Bittensor miner builder - ingest subnet configs, build competitive miners, deploy and iterate"
requires-python = ">=3.10"
dependencies = [
    "bittensor>=8.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jinja2>=3.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[project.scripts]
bt-miner-skills = "bt_miner_skills.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"
